// src/types/enums/module.enum.ts
export enum ModuleEnum {
    pos = 'pos',
    config = 'config',
    dashboard = 'dashboard',
    profile = 'profile',
    productCategories = 'product-categories',
    products = 'products',
    paymentMethods = 'payment-methods',
    clients = 'clients',
    company = 'company',
    branches = 'branches',
    roles = 'roles',
    users = 'users',
}